+++ PSManagement/Model/facturas.cs
@@ -22,8 +22,8 @@ namespace PSManagement.Model
     
         public int IdFactura { get; set; }
         public System.DateTime FechaCreacion { get; set; }
-        public double ImporteTotalSinIVa { get; set; }
-        public double ImporteTotalConIva { get; set; }
+        public float ImporteTotalSinIVa { get; set; }
+        public float ImporteTotalConIva { get; set; }
         public string TipoDePago { get; set; }
     
         [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Usage", "CA2227:CollectionPropertiesShouldBeReadOnly")]
